html/browsers/browsing-the-web/navigating-across-documents/004.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#navigating-across-documents
    steps: [3]

description: |
  This test is operationally equivalent to 003.html (navigation during unload must NOT be allowed, per https://github.com/whatwg/html/commit/2a11eef8f92b0335328890dc2cb1e281d18553eb), with the exception that 004-1.html has an `onload` handler, which sends the message "004-1" to 004.html (`messages` will be `["004-1"]`), then sets `location` to 004-2.html (as a "cross origin" resource, via `location.href.replace("http://", "http://www."`) ~100ms later.

/-->
<title>Navigation from unload whilst traversing cross-origin history</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="004-1.html"></iframe>
<script>
var t = async_test();
var iframe = document.getElementsByTagName("iframe")[0];
var messages = [];

onmessage = t.step_func(function(e) {
  messages.push(e.data);
  if (messages.length === 3) {
    assert_array_equals(messages, ["004-1", "004-2", "004-1"]);
    t.done();
    iframe.parentNode.removeChild(iframe);
  }
});
</script>
